Replace the printer-driver-gutenprint postinst with a ppd-updater trigger
Update the PPD files of all already existing print queues using this driver
when updating this package. In contrary to Gutenprint's own cups-genppdupdate,
the script of the cups package can also migrate PPD files. In our case we
migrate Foomatic/IJS PPD files to the equivalent PPDs using the CUPS Raster
driver of Gutenprint and we replace simplified PPD files by the corresponding
standard PPDs, as we discontinue the simplified PPDs in Ubuntu to reduce the
delta to Debian and to make printer model/driver listings in printer setup
tools less cluttered.

Git-Dch: Full